Add files via upload
I uploaded last week a `get_full_idx` file; however, it was slow because of a large number of for loops, which handle one case at a time. 
I rewrote this code and it now takes much less time. For large cases it will still take a few minutes to finish.
this code utilizes a for, which might be annoying for small cases. you can disable that easily; by replacing every `parfor` with `for`. no other changes required
The instructions for `get_full_index` are the same, and explained inside the file (see `help get_full_idx`)