Add files via upload
Similar to the list of variables & constraints you get when you call the `om` object (simply om). This list merges all variables (or constraints) for the same units (generators or storage)
`get_full_idx` creates a full list detailed list. The index in this list corresponds to which row in QP.x, QP.xmax and QP.xmin (or in case of linear constraints, each row in QP.A, QP.l and QP.u)

Explanation of the inputs and outputs is provided inside