Add files via upload
my personal version of `most.m` with several modifications here an there.
The main useful modification, which i'd recommend to be added to the master branch, is defining the binary variables
`u_sc` and `u_sd`, and their constraint
`u_sc` reports/denotes whether a storage device is charging up or not
`u_sd` reports/denotes whether a storage device is discharging or not

obviously: `0 <= u_sc + u_sd <= 1`. This constraint is inside the code

These variables may sound redundant, with the existence of their parents `P_sc` and `P_sd`, however,
to implement a constraint such as: storage#1 and storage#2 (on different locations) should charge up together, or discharge together
or: it is not allowed for any storage device to charge up while there exists another storage device which is discharging (selfish behavior somehow). Then you need these binary variables.
The constraint i talked about above, is not in this `most.m` file. I implemented it in one of my private files.